finished add function

+++ demo/src/Entity/GiaoDich.java
@@ -3,17 +3,18 @@ package Entity;
 import java.sql.Date;
 
 public class GiaoDich {
-    protected long maGD;
+    protected int maGD;
     protected Date ngayGD;
     protected double donGia;
     protected double dienTich;
-    protected GiaoDich(long maGD, Date ngayGD, double donGia, double dienTich){
+    protected GiaoDich(int maGD, Date ngayGD, double donGia, double dienTich){
         super();
         this.maGD = maGD;
         this.ngayGD=ngayGD;
         this.donGia=donGia;
         this.dienTich=dienTich;
     }
+    
     public double getdonGia(){
         return donGia;
     }
