+++ demo/src/controller/ThemGDController.java
@@ -1,9 +1,7 @@
 package controller;
 
 import Entity.GiaoDich;
-
 import java.sql.Date;
-
 import Entity.Dat;
 import Entity.Nha;
 import database.ThemGDDAO;
@@ -18,12 +16,12 @@ public class ThemGDController {
         themGDDAO = new ThemGDDAO();
     }
 
-    public void themGD(long maGD, Date ngayGD, double donGia, double dienTich, String loaiDat){
+    public void themGD(int maGD, Date ngayGD, double donGia, double dienTich, String loaiDat){
         Dat dat = new Dat(maGD, ngayGD, donGia, dienTich, loaiDat);
         themGD(dat);
     }
 
-    public void themGD(long maGD, Date ngayGD, double donGia, double dienTich, String diaChi, String loaiNha){
+    public void themGD(int maGD, Date ngayGD, double donGia, double dienTich, String diaChi, String loaiNha){
         Nha nha = new Nha(maGD, ngayGD, donGia, dienTich, diaChi, loaiNha);
         themGD(nha);
     }
